inlines cpp
