--- poptorch_experimental_addons/cpp/replicatedalltoallx.hpp
@@ -1,32 +1,0 @@
-// Copyright (c) 2023 Graphcore Ltd. All rights reserved.
-
-#ifndef GUARD_NEURALNET_REPLICATEDALLTOALLX_HPP
-#define GUARD_NEURALNET_REPLICATEDALLTOALLX_HPP
-
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wsign-compare"
-#pragma GCC diagnostic ignored "-Wunused-parameter"
-#include <popart/popx/op/collectives/collectivesx.hpp>
-#pragma GCC diagnostic pop
-
-namespace popart {
-namespace popx {
-
-class ReplicatedAllToAllOpx : public CollectivesBaseOpx {
-   public:
-    ReplicatedAllToAllOpx(Op*, Devicex*);
-    void grow(poplar::program::Sequence&) const final;
-    InputCreatorType getInputCreatorType(InIndex index) const final;
-    poplar::Tensor unwindTensorLayout(poplar::Tensor, InIndex, OutIndex) const final;
-    view::RegMap unwindRegion(InIndex, OutIndex) const final;
-};
-
-class ReplicatedAllToAllGradOpx : public ReplicatedAllToAllOpx {
-   public:
-    ReplicatedAllToAllGradOpx(Op*, Devicex*);
-};
-
-}  // namespace popx
-}  // namespace popart
-
-#endif
